multimon-ng: added build for building with cmake
Currently, multimon-ng builds with qmake.  However, this requires pulling in X11
packages in order to install.  Since multimon-ng is a console application the
requirement of X11 shouldn't be required.  The fix is to use cmake to build
the project.  This patch adds a CMakeLists.txt file in order to build the project.

The project was succesfully built on Mac OS (Darwin), Linux CentOS 7 with gcc and on
Microsoft Windows with Microsoft Visual Studio 2012.  The resulting program was ran
on Linux CentOS 7 and succesfully processes POCSAG512 messages.

To build without X11 support on Linux:
$ cmake -DX11_SUPPORT=OFF .
$ make

To build without Pulse Audio support on Linux:
$ cmake -DPULSE_AUDIO_SUPPORT=OFF .
$ make

Note that: X11 and Pulse Audio libraries aren't searched for on either Microsoft Windows or
Macintosh