Minor tweak to function code comparison
There appear to be two differing "standards" for how the function bit denotes the encoding of the msg.

https://www.itu.int/dms_pubrec/itu-r/rec/m/R-REC-M.584-2-199711-I!!PDF-E.pdf appears to be the original standard where 0=numeric and 3=alpha-numeric.

http://www.pa0qy.nl/pdf/PCF5001.PDF is a datasheet for the PCF5001 POCSAG decoder IC which allows for 4 differing methods depending on a separate boolean within the EEPROM of the IC whereby if it's 0, everything is decoded as numeric.  If it's 1, function code 0 is numeric; 1, 2 or 3 is alpha-numeric.

This tweak allows for decoding of both standards.  If you find you are not decoding anything useful, try forcing numeric decoding (-f numeric) to simulate the boolean  0 in the EEPROM, otherwise boolean 1 is covered by default.