Add a stub replacement for BCHCode.c
This adds a stub replacement for BCHCode.c which can be used by
distributors of binary releases of multimon-ng. If BCHCode.c is
removed then the stub is used instead and the resulting work is
compatible with the GPL. The drawback is that the FLEX decoder will
not detect and correct bit errors.

Part of #108.