Pass --ignore-length to SoX to support piped WAV data
When multimon-ng reads a WAV file it uses SoX to convert it to raw
samples. If that data is streamed via a pipe then the WAV header will
not contain the right length, so SoX will terminate early.

This adds --ignore-length to SoX's argument list, which will make it
read until the end of the input file.